refactor: minimalism audit — docs trim, CLI split, CI, branding
- Remove portfolio/generic docs; single DEPLOYMENT_GUIDE + Docker + URLs
- Short PROJECT_OVERVIEW; add docs/ARCHITECTURE, ROADMAP, README index
- Drop next_steps.md (merged into docs/ROADMAP.md)
- Move Click commands to src/cli/commands.py; slim main.py
- Add GitHub Actions pytest workflow
- Rename CommercialLLMAnalyzer → OpenRouterAnalyzer
- Align branding (Call Analytics Platform, call-analytics paths)
- Fix systemd Documentation URLs; version 5.0.2

Made-with: Cursor
