+++ .env.example
@@ -1,4 +1,4 @@
-# Шаблон переменных окружения для ASR-4.5
+# Environment template — Call Analytics Platform
 # Скопируйте в .env и заполните реальными значениями
 
 # Telegram Bot Configuration
