+++ config.example.yaml
@@ -1,4 +1,4 @@
-# ⭐ ASR-4.5 Production Configuration Template
+# Call Analytics Platform — configuration template
 #
 # ШАБЛОН БЕЗОПАСНОЙ КОНФИГУРАЦИИ
 # Скопируйте в config.yaml и заполните реальными значениями
