+++ docs/README.md
@@ -1,0 +1,20 @@
+# Documentation index
+
+Technical documentation for this repository only. Marketing material for other ScanovichAI offerings was removed from this tree to keep the OSS repo focused.
+
+| Document | Description |
+|----------|-------------|
+| [ARCHITECTURE.md](ARCHITECTURE.md) | Pipeline and components (`src/`, data flow). |
+| [ROADMAP.md](ROADMAP.md) | Planned improvements and backlog themes. |
+| [REMOTE_ASR_AND_LLM.md](REMOTE_ASR_AND_LLM.md) | Remote OpenAI-compatible LLM and ASR notes. |
+| [examples/](examples/README.md) | Synthetic sample transcript and JSON outputs. |
+
+Root-level guides:
+
+- [`../README.md`](../README.md) — quick start (English)
+- [`../README_EN.md`](../README_EN.md) — full English guide
+- [`../DEPLOYMENT_GUIDE.md`](../DEPLOYMENT_GUIDE.md) — deployment and 24/7 operations (Russian)
+- [`../PROJECT_OVERVIEW.md`](../PROJECT_OVERVIEW.md) — high-level product summary and links
+- [`../SECURITY.md`](../SECURITY.md), [`../CONTRIBUTING.md`](../CONTRIBUTING.md)
+
+Commercial services and extended positioning: [scanovich.ai](https://scanovich.ai).
