--- docs/voice-analytics-overview.md
@@ -1,212 +1,0 @@
-# 🎙️ Voice Analytics Platform Overview
-
-**Обзор решений для интеллектуального анализа голосового общения в бизнесе**
-
----
-
-## 📋 Описание решения
-
-Voice Analytics Platform - это интеллектуальная система для автоматического анализа качества обслуживания клиентов через обработку телефонных разговоров. Система помогает компаниям улучшать сервис, обучать персонал и оптимизировать бизнес-процессы.
-
-### 🎯 Целевая аудитория
-
-- **Call-центры** - оценка работы операторов, соблюдение скриптов
-- **Медицинские центры** - контроль записи на прием, консультаций
-- **Ритейл и сервис** - анализ качества консультаций
-- **Банки и финансы** - контроль продаж, техподдержки
-- **Образование** - оценка работы менеджеров по продажам
-
----
-
-## 🚀 Основные возможности
-
-### 📞 **Автоматическая транскрипция**
-- Высокоточная конвертация речи в текст
-- Поддержка различных аудиоформатов (MP3, WAV, M4A)
-- Обработка больших объемов данных (до 100K+ звонков/месяц)
-- Real-time и batch обработка
-
-### 🔍 **Анализ качества обслуживания**
-- Оценка по настраиваемым критериям (до 30+ параметров)
-- Анализ соблюдения скриптов и стандартов
-- Выявление ключевых ошибок и проблемных зон
-- Объективная оценка сотрудников
-
-### 🛡️ **Защита конфиденциальности**
-- Автоматическое маскирование персональных данных (PII)
-- Анонимизация чувствительной информации
-- Соответствие требованиям GDPR и аналогичным
-- Локальная обработка без внешних API
-
-### 📊 **Многоуровневая аналитика**
-- **Оперативные отчеты** - ежедневные/еженедельные в Telegram
-- **Детальные дашборды** - в Google Sheets с временными рядами
-- **Глубокий анализ** - CSV экспорт для Excel-анализа
-- **Управленческие insights** - рекомендации для улучшения
-
----
-
-## 🏗️ Архитектура решения
-
-### **Компоненты системы**
-
-1. **Input Layer** - прием и валидация аудиофайлов
-2. **Processing Engine** - транскрипция и анализ речи
-3. **Analytics Engine** - обработка и оценка качества
-4. **Reporting System** - генерация отчетов и дашбордов
-5. **Integration Layer** - подключение к внешним системам
-
-### **Масштабируемость**
-- **Horizontal scaling** - добавление вычислительных узлов
-- **Load balancing** - равномерное распределение нагрузки
-- **Queue management** - обработка пиковых нагрузок
-- **Resource optimization** - эффективное использование GPU/CPU
-
----
-
-## 📈 **Бизнес-метрики и эффективность**
-
-### **Производительность**
-- **37x быстрее реального времени** - обработка аудио
-- **45 секунд на звонок** - полный цикл анализа
-- **100% автоматизация** - от загрузки до отчета
-- **24/7 работа** - без остановок и простоев
-
-### **Экономическая эффективность**
-- **Экономия до $50K/год** - vs коммерческие API решения
-- **ROI 2 недели** - быстрая окупаемость инвестиций
-- **100% локально** - никаких recurring платежей
-- **Без external зависимостей** - полная независимость
-
-### **Качество результатов**
-- **Высокая точность** - профессиональный уровень транскрипции
-- **Стабильность** - 100% обработка без потерь
-- **Детализация** - 30+ критериев оценки
-- **Актуальность** - real-time обновление данных
-
----
-
-## 🔧 **Технические характеристики**
-
-### **Аппаратные требования**
-- **GPU** - NVIDIA RTX 4090/5090 (24-32GB VRAM)
-- **RAM** - 32GB+ для больших объемов
-- **Storage** - SSD 500GB+ для моделей и данных
-- **Network** - стабильное подключение для интеграций
-
-### **Программное обеспечение**
-- **OS** - Linux (Ubuntu 22.04+)
-- **Python** - 3.12+ с оптимизациями
-- **AI Models** - локальные LLM и Whisper
-- **Database** - SQLite для аналитики
-- **Web Framework** - FastAPI для API
-
-### **Интеграции**
-- **Telegram Bot** - для оперативных отчетов
-- **Google Sheets** - для дашбордов и детализации
-- **REST APIs** - для интеграции с CRM/ERP
-- **Webhooks** - для real-time уведомлений
-
----
-
-## 💼 **Примеры использования**
-
-### **Call-центр (200 операторов)**
-```
-📞 Входящие: ~5000 звонков/день
-⏱️  Обработка: 45 сек/звонок
-📊 Оценка: 25 критериев качества
-📈 Результат: +35% улучшение качества за 3 месяца
-💰 Экономия: $40K/год на обучении персонала
-```
-
-### **Медицинский центр (5 филиалов)**
-```
-🏥 Консультации: ~200 звонков/день
-🎯 Фокус: скрипты записи на прием
-📋 Критерии: 15 параметров оценки
-📊 Отчетность: ежедневные дашборды
-✨ Результат: стандартизация процесса во всех филиалах
-```
-
-### **Ритейл сеть (15 магазинов)**
-```
-🛍️ Продажи: ~800 звонков/день
-💰 Апсейл: отслеживание дополнительных продаж
-📞 Консультации: оценка качества рекомендаций
-📈 Рост: +25% среднего чека за 2 месяца
-🏆 Лучшие практики: распространены по сети
-```
-
----
-
-## 🔒 **Безопасность и compliance**
-
-### **Защита данных**
-- ✅ **PII маскирование** - автоматическое обезличивание
-- ✅ **Локальная обработка** - данные не покидают инфраструктуру
-- ✅ **Шифрование** - все данные в зашифрованном виде
-- ✅ **Access control** - строгий контроль доступа
-
-### **Соответствие стандартам**
-- ✅ **GDPR compliance** - полное соответствие
-- ✅ **Industry standards** - медицинские, финансовые нормативы
-- ✅ **Audit trails** - полная трассировка изменений
-- ✅ **Data retention** - автоматическая очистка по политикам
-
----
-
-## 🚀 **Быстрый старт**
-
-### **Фазы внедрения**
-
-1. **Анализ и планирование** (1-2 недели)
-   - Изучение бизнес-процессов
-   - Определение критериев оценки
-   - Настройка интеграций
-
-2. **Установка и настройка** (1-2 недели)
-   - Развертывание инфраструктуры
-   - Обучение моделей
-   - Тестирование на пилотных данных
-
-3. **Запуск и мониторинг** (непрерывно)
-   - Production запуск
-   - Настройка отчетности
-   - Оптимизация и развитие
-
-### **Поддержка**
-- 📚 **Документация** - полные руководства
-- 🛠️ **Техническая поддержка** - оперативная помощь
-- 📞 **Обучение** - для администраторов и менеджеров
-- 🔄 **Обновления** - регулярные улучшения
-
----
-
-## 💡 **Почему это работает**
-
-### **Технологические преимущества**
-- **Локальные AI модели** - высокая скорость и конфиденциальность
-- **Production-ready** - проверено в реальном бизнесе
-- **Масштабируемость** - от малого до enterprise
-- **Гибкость** - адаптация под любую отрасль
-
-### **Бизнес-ценность**
-- **Объективная оценка** - без субъективных мнений
-- **Автоматизация** - освобождение времени менеджеров
-- **Аналитика** - data-driven принятие решений
-- **Развитие** - рост качества сервиса и продаж
-
----
-
-## 📞 **Связаться с нами**
-
-Готовы внедрить голосовую аналитику в вашем бизнесе?
-
-📧 **Email:** iamfuyoh@gmail.com
-💬 **Telegram:** [@ScanovichAI](https://t.me/ScanovichAI)
-🌐 **Website:** [scanovich.ai](https://scanovich.ai/)
-
----
-
-*Voice Analytics Platform - это готовое решение для компаний, которые ценят качество обслуживания и стремятся к постоянному улучшению.*
