--- next_steps.md
@@ -1,89 +1,0 @@
-# Планы развития проекта ASR-4.5
-
-## Ближайшие цели (1-3 месяца)
-
-### Производительность и масштабирование
-- [ ] Оптимизация обработки больших батчей звонков
-- [ ] Поддержка нескольких GPU для параллельной обработки
-- [ ] Кэширование результатов транскрипции для повторной обработки
-- [ ] Оптимизация использования памяти для LLM-30B
-
-### Функциональность
-- [ ] Поддержка дополнительных языков (казахский, английский) в анализе качества
-- [ ] Расширение скриптов оценки (Type-C, Type-D для других отраслей)
-- [ ] Интеграция с популярными CRM системами (AmoCRM, Bitrix24)
-- [ ] Real-time анализ качества во время звонка (опционально)
-
-### Аналитика
-- [ ] Детальная аналитика по времени суток (peak hours analysis)
-- [ ] Анализ эмоций в голосе (sentiment analysis)
-- [ ] Прогнозирование результатов продаж на основе качества звонков
-- [ ] Интеграция с BI системами (Tableau, PowerBI)
-
-### Удобство использования
-- [ ] Веб-интерфейс для просмотра результатов (опционально)
-- [ ] API для интеграции с внешними системами
-- [ ] Поддержка Docker контейнеризации для легкого развертывания
-- [ ] Автоматическое обновление системы через CI/CD
-
-## Долгосрочные планы (3-12 месяцев)
-
-### Технологические улучшения
-- [ ] Миграция на более современные модели ASR (Whisper v4, когда будет доступен)
-- [ ] Эксперименты с smaller LLM для быстрой обработки (7B, 13B)
-- [ ] Поддержка других ASR моделей (Wav2Vec2, Conformer)
-- [ ] Интеграция с speech diarization для определения спикеров
-
-### Бизнес-функции
-- [ ] Автоматическое создание обучающих материалов на основе ошибок
-- [ ] Рекомендательная система для улучшения качества обслуживания
-- [ ] Интеграция с системами обучения персонала
-- [ ] Автоматическая генерация отчетов для руководства
-
-### Масштабирование
-- [ ] Поддержка multi-tenant архитектуры
-- [ ] Распределенная обработка через очередь задач (Celery, RabbitMQ)
-- [ ] Кластерная обработка для больших объемов данных
-- [ ] Поддержка облачных провайдеров (AWS, GCP, Azure)
-
-### Безопасность и соответствие
-- [ ] Поддержка GDPR compliance для работы с персональными данными
-- [ ] Аудит безопасности и penetration testing
-- [ ] Шифрование данных на уровне хранилища
-- [ ] Поддержка федерального хранения данных (для госструктур)
-
-## Исследовательские направления
-
-### AI/ML улучшения
-- [ ] Fine-tuning LLM на специфичных скриптах обслуживания
-- [ ] Использование RAG для контекстного анализа качества
-- [ ] Автоматическое определение типа звонка (продажа, поддержка, консультация)
-- [ ] Предсказание успешности звонка на основе первых секунд
-
-### Интеграции
-- [ ] Поддержка различных АТС (Asterisk, FreeSWITCH, Cisco)
-- [ ] Интеграция с системами записи звонков (CallRec, Манго)
-- [ ] Поддержка VoIP протоколов (SIP, WebRTC)
-- [ ] Обработка звонков в реальном времени через WebSocket
-
-## Приоритизация
-
-Приоритеты определяются на основе:
-- Запросов пользователей и клиентов
-- Технической сложности реализации
-- Бизнес-ценности для пользователей
-- Доступности ресурсов для разработки
-
-## Вклад в проект
-
-Если вы хотите внести вклад в развитие проекта:
-1. Создайте issue с описанием предложения
-2. Обсудите реализацию перед началом работы
-3. Следуйте стандартам кода проекта (ruff, pyright)
-4. Добавьте тесты для новой функциональности
-5. Обновите документацию
-
----
-
-**Последнее обновление:** 2025-11-04
-
