+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "call-analytics-platform"
-version = "5.0.1"
+version = "5.0.2"
 description = "End-to-end call analytics: VoIP download, Whisper transcription, quality analysis, dashboards"
 readme = "README.md"
 requires-python = ">=3.12,<3.13"
